src_data_aug/requirements.txt
absl-py==1.1.0
accelerate==0.13.0
accelerator==2022.8.4.dev1
adapter-transformers==3.1.0
aiohttp==3.8.1
aiosignal==1.2.0
altair==4.2.2
antlr4-python3-runtime==4.9.3
apex @ file:///home/zf/apex
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bayesian-optimization==1.4.3
beautifulsoup4==4.11.1
bitsandbytes==0.38.1
bitsandbytes-cuda110==0.26.0.post2
bleach==5.0.1
blinker==1.6.2
bottle==0.12.23
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
cycler==0.11.0
d2l==0.17.5
datasets==2.5.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.6
docker-pycreds==0.4.0
einops==0.5.0
entrypoints==0.4
et-xmlfile==1.1.0
evaluate==0.2.2
executing==0.8.3
faiss-gpu==1.7.2
fastjsonschema==2.15.3
filelock==3.8.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.3
futures==3.0.5
gdown==4.7.1
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.9.0
google-auth-oauthlib==0.4.6
goslate==1.5.4
grpcio==1.47.0
higher==0.2.1
huggingface-hub==0.10.0
hydra-core==1.2.0
hyperopt==0.2.7
idna==2.10
importlib-metadata==4.12.0
importlib-resources==5.8.0
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.6.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
kiwisolver==1.4.3
Levenshtein==0.21.0
lightly==1.2.22
lightly-utils==0.0.2
loralib==0.1.1
Markdown==3.3.7
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==0.8.4
more-itertools==8.14.0
msgpack==1.0.5
multidict==6.0.2
multiprocess==0.70.13
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==3.1
nlpaug==1.1.11
nltk==3.7
notebook==6.4.12
numpy==1.24.3
oauthlib==3.2.0
omegaconf==2.2.2
openpyxl==3.0.10
packaging==21.3
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.5.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.6
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.1b0
pyDeprecate==0.3.2
PyDictionary==2.0.1
PyExecJS==1.5.1
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
python-Levenshtein==0.21.0
pytorch-lightning==1.6.4
pytorch-model-summary==0.1.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
rank-bm25==0.2.2
rapidfuzz==3.0.0
ray==2.4.0
regex==2022.9.13
requests==2.25.1
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.3.5
rsa==4.8
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.2
Send2Trash==1.8.0
sentencepiece==0.1.99
sentry-sdk==1.22.1
seqeval==1.2.2
setproctitle==1.3.2
simpletransformers==0.63.11
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn==0.0
smart-open==6.3.0
smmap==5.0.0
solidity-parser==0.1.1
soupsieve==2.3.2.post1
stack-data==0.3.0
streamlit==1.22.0
synonyms==3.18.0
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX @ file:///home/conda/feedstock_root/build_artifacts/tensorboardx_1654638215170/work
terminado==0.15.0
textblob==0.17.1
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.12.0
torchmetrics==0.9.2
torchvision==0.13.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.22.2
typing_extensions==4.3.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
utils==1.0.1
validators==0.20.0
virtualenv==20.21.0
waitress==2.1.2
wandb==0.15.2
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.1
xlrd==1.2.0
XlsxWriter @ file:///home/conda/feedstock_root/build_artifacts/xlsxwriter_1645975313810/work
xxhash==3.0.0
yarl==1.7.2
zipp==3.8.0
